Agora é superclasse de Lavagem
Os métodos Abstract são aqueles que não podem ser executados na classe mãe, só nas derivadas.
Nesse caso, você só pode executar e estimar o tempo de espera para um serviço específico (Lavagem), não há como fazer isso para um objeto de Service.
